NN-5419: move getCompletedAdjudications to ManageAdjudicationsUserTokensClient
