+++ server/data/manageAdjudicationsSystemTokensClient.ts
@@ -19,9 +19,7 @@ import {
   ReportedAdjudicationResult,
   ReportedAdjudicationResultV2,
   ReportedAdjudication,
-  ReportedAdjudicationFilter,
   ReviewAdjudication,
-  allStatuses,
   ReportedAdjudicationDISFormFilter,
   ReportedAdjudicationsResult,
   allIssueStatuses,
@@ -162,28 +160,6 @@ export default class ManageAdjudicationsSystemTokensClient {
       data: payload,
     })
   }
-
-  private getCompletedAdjudications =
-    (prefix: string) =>
-    async (
-      filter: ReportedAdjudicationFilter,
-      pageRequest: ApiPageRequest
-    ): Promise<ApiPageResponse<ReportedAdjudication>> => {
-      const path =
-        `${prefix}?page=${pageRequest.number}&size=${pageRequest.size}` +
-        `${(filter.fromDate && `&startDate=${momentDateToApi(filter.fromDate)}`) || ''}` +
-        `${(filter.toDate && `&endDate=${momentDateToApi(filter.toDate)}`) || ''}` +
-        `${(filter.status && `&status=${filter.status}`) || `&status=${allStatuses}`}` +
-        `${(filter.transfersOnly && `&transfersOnly=${true}`) || ''}`
-
-      return this.restClient.get({
-        path,
-      })
-    }
-
-  getAllCompletedAdjudications = this.getCompletedAdjudications('/reported-adjudications/reports')
-
-  getYourCompletedAdjudications = this.getCompletedAdjudications('/reported-adjudications/my-reports')
 
   async getReportedAdjudicationIssueData(
     filter: ReportedAdjudicationDISFormFilter
